Move utils and fixtures around
