+++ tests/integration/test_rl_lora.py
@@ -4,12 +4,8 @@ from typing import Callable
 
 import pytest
 
-from tests.integration.conftest import (
-    ProcessResult,
-    check_number_goes_up_or_down,
-    check_number_in_range,
-    strip_escape_codes,
-)
+from tests.conftest import ProcessResult
+from tests.utils import check_number_goes_up_or_down, check_number_in_range, strip_escape_codes
 
 pytestmark = [pytest.mark.gpu, pytest.mark.slow]
